Extract app.py request-lifecycle hooks and error handlers into hooks.py / errors.py
Move the before/after/teardown request hooks into vtsearch/hooks.py and the
global JSON error handlers into vtsearch/errors.py, each exposing a
register_hooks(app) / register_error_handlers(app) that app.py calls.
Registration order is preserved.
